% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Dataset for the mixedbiastest Package}
\format{
A data frame with 97 rows and 3 variables:
\describe{
  \item{y}{Numeric response variable.}
  \item{x}{Numeric predictor variable.}
  \item{group}{Integer indicating group membership.}
}
}
\usage{
example_data
}
\description{
A dataset containing 97 observations of three variables: \code{y}, \code{x}, and \code{group}.
}
\keyword{datasets}
