% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mixedbiastest}
\alias{plot.mixedbiastest}
\title{Plot Method for Bias Diagnostic Results}
\usage{
\method{plot}{mixedbiastest}(x, ...)
}
\arguments{
\item{x}{An object of class `"mixedbiastest"`.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A \code{ggplot} object showing permutation distributions for all fixed effects.
}
\description{
Plots the permutation distributions and observed test statistics for each fixed effect.
}
