% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mixedbiastest}
\alias{print.mixedbiastest}
\title{Print Method for Bias Diagnostic Results}
\usage{
\method{print}{mixedbiastest}(x, ...)
}
\arguments{
\item{x}{An object of class `"mixedbiastest"`.}

\item{...}{Additional arguments (currently not used).}
}
\value{
No return value. This function is called for its side effects (printing the results).
}
\description{
Prints the results of the bias diagnostic in a formatted table.
}
